+++ bakis/Controllers/EmployeesController.cs
@@ -46,6 +46,137 @@ namespace bakis.Controllers
             return Ok(employee);
         }
 
+        // GET: api/Employees/5/manager
+        [HttpGet("{id}/manager")]
+        public async Task<IActionResult> GetManager([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+            string name = employee.Name + ' ' + employee.Surname;
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(name);
+        }
+
+        // GET: api/Employees/5/isActive
+        [HttpGet("{id}/isActive")]
+        public async Task<IActionResult> GetActiveParam([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            string param;
+
+            if (employee.IsActive == true)
+            {
+                param = "Aktyvus";
+            }
+            else param = "Neaktyvus";
+
+            return Ok(param);
+        }
+
+        // GET: api/Employees/5/exams
+        [HttpGet("{id}/exams")]
+        public async Task<IActionResult> GetEmployeeExams([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            var exams = _context.EmployeeExams.Where(l => l.EmployeeId == id);
+
+            return Ok(exams);
+        }
+
+        // GET: api/Employees/5/certificates
+        [HttpGet("{id}/certificates")]
+        public async Task<IActionResult> GetEmployeeCertificates([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            var cert = _context.EmployeeCertificates.Where(l => l.EmployeeId == id);
+
+            return Ok(cert);
+        }
+
+        // GET: api/Employees/5/duties
+        [HttpGet("{id}/duties")]
+        public async Task<IActionResult> GetEmployeeDuties([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            var duties = _context.EmployeeDuties.Where(l => l.EmployeeId == id);
+
+            return Ok(duties);
+        }
+
+        // GET: api/Employees/5/competencies
+        [HttpGet("{id}/competencies")]
+        public async Task<IActionResult> GetEmployeeCompetencies([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
+            {
+                return NotFound();
+            }
+
+            var competencies = _context.EmployeeCompetencies.Where(l => l.EmployeeId == id);
+
+            return Ok(competencies);
+        }
+
         // PUT: api/Employees/5
         [HttpPut("{id}")]
         public async Task<IActionResult> PutEmployee([FromRoute] int id, [FromBody] Employee employee)
@@ -109,6 +240,18 @@ namespace bakis.Controllers
             if (employee == null)
             {
                 return NotFound();
+            }
+
+            var empl = _context.Employees.Where(l => l.FkEmployeeId == id).Select(l => l.EmployeeId).FirstOrDefault().ToString(); //patikrinti
+            var salary = _context.Salaries.Where(l => l.EmployeeId == id).Select(l => l.SalaryId).FirstOrDefault().ToString();
+            var employeeCertificate = _context.EmployeeCertificates.Where(l => l.EmployeeId == id).Select(l => l.EmployeeCertificateId).FirstOrDefault().ToString();
+            var employeeExam = _context.EmployeeExams.Where(l => l.EmployeeId == id).Select(l => l.EmployeeExamId).FirstOrDefault().ToString();
+            var employeeCompetency = _context.EmployeeCompetencies.Where(l => l.EmployeeId == id).Select(l => l.EmployeeCompetencyId).FirstOrDefault().ToString();
+            var employeeDuties = _context.EmployeeDuties.Where(l => l.EmployeeId == id).Select(l => l.EmployeeDutyId).FirstOrDefault().ToString();
+
+            if (empl != "0" || salary != "0" || employeeCertificate != "0" || employeeExam != "0" || employeeCompetency != "0" || employeeDuties != "0")
+            {
+                return BadRequest("Negalima ištrinti šio darbuotojo, nes jis turi susijusių įrašų.");
             }
 
             _context.Employees.Remove(employee);
